added /discover api where we can fetch all users except the current user, and also get and set their topTracks in the User model.
added /discover api where we can fetch all users except the current user, and also get and set their topTracks in the User model.

Also added findBySpotifyIdNot  in UserProfileRepository so that spring jpa filters from the db itself, we dont have to add different logic to fetch all users except current.

This is for the matching page where we display the recommendations of users to match with

At the moment we do not have any logic to display them, we are simply displaying all Users